They are in seconds ;)
Revert "changed delta to ms"

This reverts commit a259693084e1913974472efbe430094c5a297765.

+++ src/main/java/org/getspout/vanilla/entity/living/hostile/Creeper.java
@@ -28,7 +28,7 @@ public class Creeper extends Monster {
 	}
 
 	@Override
-	public void onTick(long dt) {
+	public void onTick(float dt) {
 		super.onTick(dt);
 	}
 }
