+++ src/main/java/org/getspout/vanilla/entity/living/hostile/Skeleton.java
@@ -30,7 +30,7 @@ public class Skeleton extends Monster implements HostileEntity, Land {
 	}
 
 	@Override
-	public void onTick(long dt) {
+	public void onTick(float dt) {
 		super.onTick(dt);
 	}
 }
