+++ src/main/java/org/getspout/vanilla/entity/living/hostile/Slime.java
@@ -30,7 +30,7 @@ public class Slime extends Living implements HostileEntity, Land {
 	}
 
 	@Override
-	public void onTick(long dt) {
+	public void onTick(float dt) {
 		super.onTick(dt);
 	}
 }
