+++ src/main/java/org/getspout/vanilla/entity/living/passive/Sheep.java
@@ -29,7 +29,7 @@ public class Sheep extends AnimalEntity implements PassiveEntity {
 	}
 
 	@Override
-	public void onTick(long dt) {
+	public void onTick(float dt) {
 		super.onTick(dt);
 	}
 }
